// src/app/interfaces/game.ts
export interface Game {
  id: number;
  age_ratings: number[];
  alternative_names: number[];
  category: number;
  collection: number;
  cover: number;
  created_at: number;
  external_games: number[];
  first_release_date: number;
  follows: number;
  game_modes: number[];
  genres: number[];
  involved_companies: number[];
  keywords: number[];
  name: string;
  platforms: number[];
  player_perspectives: number[];
  rating: number;
  rating_count: number;
  release_dates: number[];
  screenshots: number[];
  similar_games: number[];
  slug: string;
  storyline: string;
  summary: string;
  tags: number[];
  total_rating: number;
  total_rating_count: number;
  updated_at: number;
  url: string;
  websites: number[];
  checksum: string;
}